# Conflicts:
#	Source/Applications/Desktop/Vodovoz/Dialogs/DocumentDialogs/ShiftChangeWarehouseDocumentDlg.cs
#	Source/Applications/Desktop/Vodovoz/JournalColumnsConfigs/JournalsColumnsConfigs.cs
#	Source/Applications/Desktop/Vodovoz/MainWindow.cs
#	Source/Applications/Desktop/Vodovoz/Vodovoz.csproj
#	Source/Applications/Desktop/Vodovoz/gtk-gui/MainWindow.cs
#	Source/Applications/Desktop/Vodovoz/gtk-gui/gui.stetic
#	Source/Applications/Desktop/VodovozViewModels/VodovozViewModels.csproj

+++ Source/Libraries/Core/Business/VodovozBusiness/EntityRepositories/Store/IWarehouseRepository.cs
@@ -14,7 +14,7 @@ namespace Vodovoz.EntityRepositories.Store
 		IEnumerable<NomanclatureStockNode> GetWarehouseNomenclatureStock(IUnitOfWork uow, int warehouseId, IEnumerable<int> nomenclatureIds);
 		IEnumerable<Nomenclature> GetDiscrepancyNomenclatures(IUnitOfWork uow, int warehouseId);
 		bool WarehouseByMovementDocumentsNotificationsSubdivisionExists(IUnitOfWork uow, int subdivisionId);
-		decimal GetTotalShippedKgByWarehousesAndProductGroups(
+		int GetTotalShippedKgByWarehousesAndProductGroups(
 			IUnitOfWork uow, DateTime dateFrom, DateTime dateTo, IEnumerable<int> productGroupsIds, IEnumerable<int> warehousesIds);
 	}
